@@ -1,6 +1,6 @@
 module github.com/faabiosr/cachego
 
-go 1.18
+go 1.19
 
 require (
 	github.com/bradfitz/gomemcache v0.0.0-20230905024940-24af94b03874
